Correctly apply tag concurrency limits
To do this, we need to paginate through runs in storage_id order so I've
added an ascending argument to get_runs. I also fixed a bug with how
cursor and ascending interact - we have to make a different comparison
for the cursor depending on if we're ascending or descending.