[dagster-databricks] Fix polling code to handle null API response (#18445)
## Summary & Motivation

The Databricks API can sometimes return a null for
`RunState.life_cycle_state`. Our best hypothesis is that this occurs if
a job is queued, which will happen when one is over the concurrency
limit. See:
https://dagsterlabs.slack.com/archives/C05HE0RG8SU/p1701401715458069

This PR changes our polling code to assume that if a `life_cycle_state`
is null, we should keep polling. Previously it would crash.

## How I Tested These Changes

BK