[docs][pythonic resources] Tweak pythonic resources docs (#18398)
## Summary

Updates the Pythonic resources docs with a few small tweaks.

The main, potentially controversial change is to add the
`ResourceDependency` wrapper to the resource-resource dependency
example. While this does complicate the code a bit, I think this is
valuable for a few reasons:

- Explicitly differentiates resource dependencies from configuration
schema. Right now under the hood, this is inferred based on whether the
annotated type is a `ConfigurableResource` or not. This behavior can
create confusion for users, and at a glance it's hard to tell just from
the set of attributes which are supposed to represent config and which
resources:

  ```python
  class MyResource(ConfigurableResource):
      # here, we have no idea whether APICredentials is a resource or a
      # Config subclass without going to its definition
# depending on which it is, it's treated very differently when
constructing
      # & initializing MyResource
      credentials: APICredentials
  ```

- Keeps behavior consistent between Pythonic resources & those which
yield another object. Any dependency on old-style resources which do not
subclass `ConfigurableResource` must already be annotated with
`ResourceDependency`. If users override `create_resource` on a
`ConfigurableResource` to yield a plain Python object, they'll also have
to use `ResourceDependency`.


   ```python
   class MyResource(ConfigurableResource):
       ...

   class MyPopoResource(ConfigurableResource[str]):
       ...

       def create_resource(context) -> str:
           return "foo"


   @resource
   def my_resource(context) -> int:
       ...


   class MyDependentResource(ConfigurableResource):
       # no need for ResourceDependency
       my_resource: MyResource

       # needs ResourceDependency
       my_popo_resource: ResourceDependency[str]
       my_old_resource: ResourceDependency[str]


   MyDependentResource(
       my_resource=MyResource(...),
       my_popo_resource=MyPopoResource(...),
       my_old_resource=my_resource
   )
   ```