[ui] Repair scroll behavior on dynamic virtualized lists (#18564)
## Summary & Motivation

Resolves #17892.

Repair a bug in virtualized dynamic-height lists in which the subpixel
row height is causing an extra scrollbar to appear, specifically in the
code locations list. The row heights are rounded for the summed
container height, which can lead to the container being just a bit
shorter than the total rendered row height, leading to an unwanted
scrollbar.

Worse, the scrollbar sometimes broke the grid layout, causing the header
column display and row column displays to become misaligned.

In this case, most notably, I needed to remove the `overflow: hidden`
wrapper div. This way, even though the container may have a height that
is less than the actual rendered row height sum, no scrollbar appears.

While working on this, I upgraded to v3.0.1 of the tanstack virtualizer
lib, which required an update to the API usage. Specifically,
`measureElement` is now on the `virtualizer` object, and the individual
rows need a `data-key` prop.

## How I Tested These Changes

View code locations list, with the `count` value restricted to a smaller
number. Verify that the unwanted scrollbar does not appear, even though
the container height is less than the dynamically rendered row height
sum.

View the side nav in the global asset graph, verify that it still
renders and behaves correctly.