[ui] Require Node 20, upgrade Yarn (#18639)
## Summary & Motivation

We've run into some issues with the dev build when using Node 18 (but
not 16 or 20) where request proxying fails for mysterious reasons. The
problem goes away with 20, and now that Vercel supports it (i.e. for our
storybook deployments), we can require it explicitly.

Also upgraded Yarn to the latest stable.

Corresponds to https://github.com/dagster-io/internal/pull/6530.

## How I Tested These Changes

BK, Vercel preview deployments.