Make all postgres connection CMs transactional (#18584)
## Summary & Motivation
We probably want all of these to happen in an explicit transaction.

Ran into this when implementing some of the concurrency changes which we
explicitly want to happen in a transaction. We're currently in a weird
state where we have explicit transactions in MySQL / sqlite, but
autocommit behavior in Postgres, so the generic sql-based
implementations are ambiguous.

Depending on the version of sqlalchemy, we either allow or error on
nested transactions. Cloud is on sqlalchemy 1.4 for various reasons
(which allows nested transactions), but if we ever upgrade to 2.0, we'll
want to get rid of the nested transactions.

## How I Tested These Changes
BK