Add an env-var-controlled offset to the cursor used by asset backfills (#18671)
Summary:
This is another potential workaround for out-of-order race conditions
with event insertions into postgres. We can take advantage of the fact
that asset backfills can process the same events more than once without
issue to add a fixed offset to the queries that. That should ensure that
even if an event slips through the cracks on the first iteration,
subsequent iterations will still pick it up via the offset.

## Summary & Motivation

## How I Tested These Changes