Enable serializing partitions subset on materialization planned event (#18418)
This PR enables serializing a partitions subset on the asset
materialization planned event. It adds an optional `partitions_subset`
field that parallels the existing `partition` field on
`AssetMaterializationPlannedData`.

Since we cannot query for partitions subsets events by partition key,
the partitions subset is not mirrored to any columns, it only exists
within the event body.