[ui] Offer to materialize changed and missing in more contexts
When we revamped the data loading in the global asset graph, we limited the “materialize changed and missing” to specific contexts (#18198).

I think that it makes sense to show this option whenever more than one asset is in scope, whether you’re operating on an all assets in view or on a selection.

Previously, we disabled this option when “scope=selection” because there was no confirmation modal and it was confusing whether it would use your selection or not. Now that we have a nice modal with checkboxes for revising + confirming the subset that will materialize, it’s no longer confusing.

I audited all the callsites where we use the LaunchAssetExecutionButton, and the only places we don’t want to show this option is on the individual asset view. To make sure this doesn’t get left out again, I inverted the option from “showChangedAndMissingOption” (show option) to “scopeAlwaysSingleAsset” (hide option) and set it in just that one place.