[ui] Fix layout for global banner (#18817)
## Summary & Motivation

The global banner some users are seeing in Cloud is breaking the layout.
This is because the banner pushes the main content down below the edge
of the viewport.

To resolve this, modify the core layout component to accommodate the
banner by passing it as a prop.

## How I Tested These Changes

Force a banner to appear on the page. Click around throughout the app,
verify that pages render correctly and can be scrolled correctly as
well, including pages with virtualized tables and DAG views.