[ui] Remove non-SDAs from global search (#18824)
## Summary & Motivation

Resolves #18669.

There is currently an issue where deleted SDAs are still appearing in
search, because we can't currently know whether a deleted SDA is any
different from a non-SDA.

To resolve this, just don't show non-SDA assets in global search. Thus,
if the definition is deleted, it will no longer appear in search, though
it will still appear in the asset catalog.

## How I Tested These Changes

Search for a non-SDA. Verify that it doesn't appear in global search.