[ui] Asset partitions: Fix invalid filter state scrollToIndex (#18772)
## Summary & Motivation

There is currently a JS error being thrown on a partitioned asset view
when attempting to view a specific partition with filters selected that
do not match the partition. We try to scroll to that partition but fail
because it doesn't exist.

Check for the existing of the partition key before trying to scroll to
it.

## How I Tested These Changes

View a partitioned asset, select a partition, then unselect the filters
that match its status. Verify that the page doesn't error. Reload the
page with those params, verify that the page shows the selected
partition and the partition list, but does not error.