Manually instrument workspace jobs root (#18843)
## Summary & Motivation
As titled, we create a trace for the WorkspaceJobsRoot component. I
might revisit some of the semantics here in the future, maybe instead of
calling this a trace I might rename the API to call it a segment since
really the trace this is going to be part of is a page load trace which
will encapsulate this segment.

## How I Tested These Changes

console.log'd the "trace" finished when the data loaded.