[ui] Sensor page branching for AMP (#18774)
## Summary & Motivation

Split the rendering of the Sensor page to handle both asset auto-materialization sensors and standard sensors.

For auto-materialization sensors, repurpose the timeline and table components used in the current Overview auto-materialization page.

<img width="1220" alt="Screenshot 2023-12-15 at 4 48 19 PM" src="https://github.com/dagster-io/dagster/assets/2823852/20f73491-ee70-4981-861e-51724d31df89">

## How I Tested These Changes

With `use_automation_policy_sensors: true`, load app and navigate to Sensors overview.

View auto-materializing sensor, verify that the timeline and evaluations table load correctly. Click an evaluation in the timeline, verify that it is the auto-materialization tick dialog. Switch from evaluations to runs, verify that the query retrieves the correct list of runs for the sensor.

View a standard sensor, verify that behavior and rendering is unchanged from the current UI.