Switch dagster_webserver and dagster_gcp_pyspark to hyphens (#18902)
Open questions:

- ~~why did tests start failing this weekend? (best guess is setuptools
69.03 on Dec 23rd)~~
- ~~why does this fix the failures?~~
- would this be a breaking change? I'm hopeful that it's not, because
the package is already published as dagster-webserver
https://pypi.org/project/dagster-webserver/

Edit: the first two are answered by
https://github.com/pypa/setuptools/issues/2522#issuecomment-1856091483:
setuptools change how they handle hyphens vs underscores

This makes the two required changes to get the build green (these two
are special because of the underscore packages, they're the ones that
are relied on in tests). In general we should make it a pattern to make
package names use hyphens and match their directory name to avoid
confusion.