Respect sensor timeout arg from proxy server (#18849)
When the proxy server calls through timeout arg to the client, it does
so without providing a timeout arg. This means that the default grpc
timeout gets used for the get_response call. This PR changes to
intercept the timeout param from the sensor request args, and pass it
through to the get_response call.

I think the existing test where we decrease the timeout can be used for
also expanding it to test this behavior.