[ui] Reuse RunConfigDialog for Runs list (#18914)
## Summary & Motivation

Resolves #18855.

Use the run config dialog from the Run page on the Runs list. It
occupies more of the viewport, which is better for runs with large
config yaml content.

## How I Tested These Changes

View a run, click to view the tags/config. Verify success.

View Runs list, use action menu to view config. Verify that the large
dialog appears properly.