[RFC] Add metrics retrieval to grpc server (#18721)
Introduces a potential API for metrics retrieval from the GRPC server. 
- We retrieve metrics about the currently running requests of each type
for a grpc server. This is implemented through a `retrieve_metrics`
decorator. The decorator manages lifecycle and stores metadata about
requests in a global threadsafe-access dictionary.
- Methods decorated with `retrieve_metrics` are added to a registrar,
which allows us to enforce that all new methods have the decorator.
- The Ping API has been repurposed to snapshot these metrics.
- Adds a test demonstrating metrics retrieval of a running sensor.
- Adds a test which could be used to enforce that all GRPC methods are
instrumented with this decorator. I'll leave the actual enforcement to a
future PR.

Using a decorator here keeps the implementation contained / doesn't
muddy up the actual methods, while still allowing us to intercept
arguments / retrieve more interesting information than just "a request
of this type is running".

Viewing slices of grpc server utilization over time like this can be
incredibly powerful. It could power a view, for example, of the number
of running sensors over the total number of workers, and which sensors
are actually being run in each of those slices.

There's definitely some potential spiciness with the global dictionary +
lock shenanigans, but I think it stays pretty constrained. We're going
to have to have some sort of global data structure to consolidate
information regardless, I think.

**Changes from round 1 of comments**
- Record max_workers from the threadpool
- Make metrics retrieval opt-in
- Only record count per request type