Log materialization planned event for run with partition range (#18305)
This PR enables the UI to correctly display failed/in progress status
for single-run backfills, in cloud.

This includes two major changes:

### 1. Log a materialization planned event containing a partitions
subset for single-run backfills upon run creation

If all of the below conditions apply, we log an partitions subset
planned event for each asset:
- Run targets a partition range
- Asset is partitioned (available on execution plan snapshot)

Otherwise, we fallback to the status quo behavior:
1. If run has a single partition, log a planned event w/ a partition key
for each asset
2. If run targets a partition range, log a planned event w/
`partition=None`. We still log a planned event in this case because in
certain places (i.e. asset backfills) we query planned events to see
which assets are planned to execute.

### 2. Threads an `assets_job_partitions_def` through the run creation
callsites

We need access to the partitions definition of the assets/job in order
to build the target partitions subset from the partition range. The
alternative would be to serialize a list of targeted partition keys on
the execution plan snapshot for partition ranged runs, which bloats the
snapshot.

There are two different ways runs are created from jobs:
1. The execution plan is built from the code location
2. Testing contexts; the job definition is passed into
`execute_in_process` and the execution plan is created from there

In the first case, the code location holds the external repository data
in memory. We can then fetch the external partitions definition of the
job, and thread this into `create_run`. We could also fetch the
partitions definition per-asset from the code location, but that feels
extraneous given that we know (1) the partitions definition on the job
and (2) whether each asset is partitioned or not.

In the second case, we don't have access to the code location, so
instead this PR constructs the external partitions definition data from
the job def and threads that into `create_run`.

### Testing
Tested locally on cloud and OSS.

Corresponding internal PR:
https://github.com/dagster-io/internal/pull/7615