[ui] Memoize timestamp rendering (#18930)
## Summary & Motivation

While doing a bit of perf profiling on the Sensor page, I noticed that
timestamp rendering in the evaluations table is one of the slowest parts
of the table. This can add up when there are a lot of timestamps
rendered on the page, and when the table is being re-rendered often as
polled data comes in.

Date formatting can be a bit slow in general, and since these are static
values with consistent configuration, we can be a bit more proactive
about avoiding extra work.

This PR adds memoization for the formatted date/time string for a given
timestamp and config.

## How I Tested These Changes

On a running automaterialization sensor, note that the `Timestamp`
component is one of the slowest parts of the evaluation table. Then,
with the change in place, verify that the `Timestamp` component
re-render time is nearly eliminated.