[ui] Fix partition status dots wrapping to next line on Asset > Partitions (#18694)
Using a grid here is a bit tricky because more than one of the statuses
can apply to a partition in a multi-dimensional asset. I think our
options are to wrap the dots in a `<Box flex gap={4}>` or convert this
parent back to a flexbox, opted for the latter.

## How I Tested These Changes

I simulated a long title and made sure it truncates still and that the
tooltip still works correctly. Also went through all the other `display:
'grid'` occurrences in the project and verified there aren't any others
with a varying # of children!

<img width="364" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/6fddae9b-6a6f-4b46-9d7c-41890b345362">

Co-authored-by: bengotow <bgotow@elementl.com>