[ui] Replace broken “Open in launchpad” links for asset jobs (#18695)
## Summary & Motivation

Fixes #16639 

The top right of the Run page has always shown an "Open in Launchpad"
button. There are two big problems:

- The launchpad / playground page does not exist for asset jobs and this
opens it anyway
- The launchpad for a hidden asset job launches the entire asset graph,
not the asset selection of the run you're viewing

This PR replaces this button (and the corresponding RunActionsMenu item)
with three different actions based on the job type:

- For op jobs, it remains "Open in Launchpad" (Launchpad is capitalized,
which is a bit odd, but we do it consistently everywhere as if it were a
proper noun, so I left it that way!)

- For hidden asset jobs, it is "View asset graph" and links to the
global graph view of the subset of assets you launched

- For proper asset jobs, it is "View job" and links to the main graph
page of the job (since asset jobs have no launchpad)
 

<img width="1422" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/8f68d133-c2f2-45bb-b2d2-5d689903d609">

## How I Tested These Changes

I tested the different disabled states (it is disabled for op jobs if
you do not have re-execute permission) by manually inserting true/false
values in the data to verify the logic was correct. Also tested with the
three different kinds of jobs above and verified the tooltips + run
actions menu work. There is also a jest test for this and I verified it
still passes.

---------

Co-authored-by: bengotow <bgotow@elementl.com>