Expose trace data via global for cypress  (#18962)
## Summary & Motivation

Currently Cypress isn't receiving the custom performance events we're
emitting because cypress tests only run after the page is fully loaded
which means our listener for the event doesn't get setup until after the
event fires. To allow cypress to get the traces we added them to a
global buffer. I don't expect the memory of this buffer to be an issue
for now but down the line I can add some gating to make it so we only
store this data if we're running cypress.

## How I Tested These Changes

dogfood