flatten & and | asset selections (#18912)
## Summary & Motivation

Prior to this PR, `&`-ing asset selections a, b, and c via `a & b & c`
would result in `AndAssetSelection(AndAssetSelection(a, b), c)`.

With this PR, it results in `AndAssetSelection([a, b, c])`.

Advantages:
- The serialized representation takes up less space
- Its easier to generate a legible string representation, to display in
the UI

Note that this breaks compatibility with previously serialized
AndAssetSelection and OrAssetSelections. But we haven't released since
we whitelisted those for serdes. So we need to merge this before we
release.

## How I Tested These Changes