[ui] Remove the “experimental asset graph” and “disable caching” feature flags (#18991)
## Summary & Motivation

The experimental asset graph is here! This PR removes the feature flag
and enables it all the time. I also converted the "disable caching"
option to a constant you can change internally for development, since
users shouldn't need to disable it anymore.

## How I Tested These Changes

Ran Dagster with localstorage erased, verified that I see the new asset
graph experience and the layout is cached.

Co-authored-by: bengotow <bgotow@elementl.com>