add base concurrency limit initialize method to event log storage (#18603)
## Summary & Motivation
This changes the behavior for non-configured concurrency keys to default
to a configurable concurrency limit default, if you have run `dagster
instance migrate`.

The behavior in the unmigrated state is the same as the status quo,
which is default open (e.g. not limited).

Right now, this default limit can be configured in the instance settings
(`concurrency` > `default_op_concurrency_limit`) in `dagster.yaml` (or
deployment settings in cloud).

I also considered storing this default in the DB, in the limits table,
for a special hard-coded concurrency key (i.e.
`DEFAULT_CONCURRENCY_KEY`), but opted for this instead.

Not included in this diff: figuring out how to coalesce some different
concurrency related settings under this new settings key (e.g. the run
monitoring setting `free_slots_after_run_end_seconds`). This is mostly
because reconciling the settings may be tricky if they don't have run
monitoring enabled, etc.. Also, unclear if we should coalesce the run
concurrency config into this concurrency section.


## How I Tested These Changes
BK