[ui] Webpack analyzer (#18996)
## Summary & Motivation

Add the `StatsWriterPlugin` to the app-oss next.js build config so that we can produce webpack stats for analysis. Based on https://relative-ci.com/documentation/guides/webpack-stats/next.js.

Not sure if there's a nice way to make it only output stats when using a CLI arg or something.

## How I Tested These Changes

`yarn build`, then `yarn analyze`. Verify that the webpack analyzer opens with the build stats.