Sensor daemon threadpool refactor (#18978)
Retain a reference to the threadpools we use in the sensor daemon, on
the daemon object itself. This allows us to introspect information about
the threadpool when control is yielded from the core loop.

Also makes some changes to the threadpool subclass we use to gate
retrieval of internals on the threadpool.