[daggy u] Fix iframed snippet copy behavior (#19069)
## Summary & Motivation

We received a report from a Dagster University user that the button to
copy code snippets is failing in Chrome, due to a permissions limitation
on the Clipboard API. I'm not sure if this has been broken for a while,
but I was able to repro the issue.

Work around this by using the old fashioned `execCommand` behavior to
write to the clipboard without the Clipboard API.

## How I Tested These Changes

yarn build, yarn start. Load an iframe with no `allow` values set, with
a course page loaded:

```
<html>
  <body>
    <iframe src="http://localhost:3000/dagster-essentials/lesson-3/defining-your-first-asset" />
  </body>
</html>
```

(In comparison, the Thinkific iframe allows a handful of features, but
similarly does not allow any Clipboard access.)

Scroll to a code snippet, verify that I can successfully copy the
snippet using the clipboard button. Verify also that newlines and
whitespace are correct.