Carbon copy of Automaterialize policy page components (#19074)
## Summary & Motivation

In case something goes wrong with the new AMP migration we want a way to
easily switch back to the old AMP page. To make that easy this PR clones
the existing AMP page into a new folder: "AutoMaterializePolicyPageOld"
which we will delete once we finalize the migration. I opted for this
instead of creating a "New" folder to preserve the existing git history
for after the migration.

## How I Tested These Changes
yarn build