Compress cursor for sensor (#19025)
## Summary & Motivation

This updates how we store the cursor on the SensorInstigatorData object.
Previously, we would just store the raw string, which is serialized in a
somewhat weird custom way.

In the near future, we want to migrate the cursor to a more traditional
whitelisted NamedTuple. Having some SensorInstigatorData objects contain
info serialized one way, and other objects containing data serialized in
a completely separate way would make codepaths more complicated than
desired.

Here, we create a thin NamedTuple wrapper to store the serialized
string, such that all information stored in this field will be a
NamedTuple. On top of that, we zlib-compress / b64 encode this
information to cut down on the total storage size.


## How I Tested These Changes