Retrieve more granular resource utilization metrics (#19078)
Previously, we were calculating utilization metrics at the container
level rather than in whichever callsite is using this information. I
think we want to move to just retrieving as granular metrics as
possible, since it's possible for any of this information to be present
or not present. Then, we can make more granular decisions about how to
use this information in the callsite.

I've added cpu_quota_us and cpu_period_fs, which you can learn more
about
[here](https://unix.stackexchange.com/questions/417506/what-is-the-effect-of-setting-cpu-cpu-quota-us-in-cpu-cgroup).
Essentially allows us to retrieve number of virtualized cpus, as well as
cpu usage under certain scenarios.

This actually makes a lot of the logic and testing simpler.