[1.6] feat(dbt)!: customize `backfill_policy` default using `partitions_def` (#17822)
## Summary & Motivation
Addresses a comment from
https://github.com/dagster-io/dagster/issues/17778#issuecomment-1801268798.

We can have a more sensible default for `backfill_policy` by using
`partitions_def`. If the partitions definition is a time window, then we
should default to a single-run backfill policy. If the partition
definition is static, then we should default to a multi-run backfill
policy.

If the `backfill_policy` is explicitly defined, ignore these defaults.

## How I Tested These Changes
pytest