Fix data version bug with corrupted state
In `DataProvenance` objects the `input_data_versions` and
`input_storage_ids` should always have the same set of keys. Code is
written based on this assumption. However, we've observed one instance
where that wasn't the case, resulting in an error. It's as yet unclear
how this happened.

This PR changes our key check in `input_data_versions` (implicitly used
to check both dicts) to instead explicitly check both dicts. The
corrupted state where a key is present in only one of these dicts is
handled the same as if the key is present in neither, that is, as if the
key is a new dependency not used in the prior materialization.