[ui] Fix SVG export option on the DAG (#19116)
## Summary & Motivation

This PR fixes
#[18651](https://github.com/dagster-io/dagster/issues/18651). There was
a forgotten todo in this code and it was still trying to use
ReactDOM.render.

It's not possible to ReactDOM.render anymore, and even upgrading to the
new root.render doesn't work because we now use ApolloClient + some
other contexts within the DAG (for the context menus). The asset DAG
can't be rendered into a freestanding React root.

This changes the implementation so that it re-renders the actual DAG,
clones the DOM nodes and exports that.

## How I Tested These Changes

Exported graphs and viewed the SVGs. Also verified the DOM is restored +
intact after exporting, and exporting over and over again works and
generates the same SVG.

Co-authored-by: bengotow <bgotow@elementl.com>