[asset backfills] Handle UI error when unpartitioned root becomes partitioned (#19121)
A graphQL error occurs when loading the asset backfill page if an asset
backfill is created with an unpartitioned root asset but the asset is
later changed to be partitioned. The `AssetBackfillData` attempts to
fetch the root target partitions subset for the root asset, but this
causes a key error as the backfill did not target a partitions subset
for this asset.

This PR fixes this by checking that an asset was partitioned at storage
time in order to consider it as a partitioned root.