[asset backfill] [non-urgent] Terminate "canceling" backfill when all runs complete  (#18692)
Certain asset backfills that hit recent race condition issues were stuck
in a `CANCELING` state. In the cancellation iteration, we wait until all
partitions are marked as materialized or failed in order to enter a
canceled state. When partitions are hanging (i.e. truly materialized,
but skipped in the iteration), this causes these backfills to hang
forever.

This PR checks if all partitions are complete (materialized/failed),
marking the backfill as failed if there are no in-progress runs.