[docs build] fix how changed files are listed (#19139)
## Summary & Motivation

Fix a recurring github actions error, due to how the commits to compare
are selected. The github action often failed because the proposed base
sha of the branch does not actually match the commit it was forked from
but the HEAD of the base ref, which will not be part of the commits
fetched by actions/checkout with the set fetch_depth unless the branch
has been rebased to that HEAD just prior to the CI execution.

See #19118. 

## How I Tested These Changes

Added a temporary bogus doc change and tested expected result in this
build:
https://github.com/dagster-io/dagster/actions/runs/7477417883/job/20350042851?pr=19139