[ui] Remove dark prop on MetadataTable (#19145)
## Summary & Motivation

The `dark` prop on `MetadataTable` is used on the log row timestamp tooltip, but as far as I can tell, it's not really needed -- the default text colors seem fine here. We applied the wrong text color here during theming, and the keys are currently invisible to people using light mode.

I think maybe the key color was slightly different from the value color before theming, but I'm not sure we really need it.

## How I Tested These Changes

In light mode, view run logs, hover on timestmap. Verify that the elapsed time label is visible. Switch to dark mode, verify same.