Typed GRPC server metrics (#19126)
This adds some typing to the grpc server metrics collection code paths,
and changes up the format to separate out different kinds of metrics
more coherently. The idea is that we can reuse these classes in multiple
places if they are scoped enough. To describe what's added:
- RequestUtilizationMetrics (to measure the throughput of a dagster
system component, usually in the form of a threadpool, but kept language
in terms of requests in case there are non-threadpool use cases we would
like to cover.
- GrpcApiMetrics (to measure various statistics about current request
throughput of a particular grpc api)
These are encapsulated by `DagsterCodeServerUtilizationMetrics`, which
contains RequestUtilizationMetrics(throughput of the threadpool),
ContainerUtilizationMetrics (kernel-level mem and cpu info) and a
dictionary mapping a grpc API name to an instance of GrpcApiMetrics
(which right now is just tracking the total number of requests of that
type being concurrently served.